feat(log-manager): add custom log levels
Adds methods addCustomLevel(name, value) and removeCustomLevel(name)
to LogManager. For testing purposes, but also of general use, adds
clearAppenders() and getAppenders().  Adding a custom level adds a
method of the corresponding name to the Logger interface.  This change
is needed to enable the logger to replace console for existing functions
like trace, time/timeEnd, etc, as well as to support custom behavior
for additional logging use cases. In my case, I need an exception()
method to format passed error stacks to submit to an error reporting
service.

There should be no breaking changes. Closes issue #33.